Test application registration.
Registration is initiated by the application performing a `POST` against
the `/register` end point of the Compassion Local API. The `POST`
payload includes the postback URL of the application.

The `POST` also includes the name of an island which is supposed to be
unique for an application. You can use the same Compassion instance for
multiple applications giving each application a unique name. Here we use
the NPM module name and if we stick to that as a convention we will
probably not have any collisions.

The `POST` also includes a unique id for this application instance. Each
application instance is a participant in the collective application. It
is up to the user to ensure that this id is unique for each instance. It
does not have to be unique for each run, though. It can be reused on
crash-restart, so you can use something unique to the particular
instance like a Kubernetes pod id.

Hmm… What really needs to be unique is the combination of island and id,
so if you're using the Kubernetes pod id you can use it with different
applications that are sharing a local Compassion instance.

Registration is where you register the post backs you're interested in,
but you always have to have a `register` post back in order to get your
authentication token. You need this token when you make a call to the
Compassion. If you get any error response, your application should
crash-restart.

Also, you can provide a token that will be used to call your end points
so that if you crash-restart so fast that Compassion doesn't see you go
missing, you can reject postbacks which will cause Compassion to remove
its entry for the application making it ready to accept a new instance.