Arrival and acclimation.
All applications are going to use the atomic log to make changes to
common set of data. If they all start with the same data and then make
the same changes in the same order, then after each change they will
have the same data that all their peers had when they made the
change.

This is called a replicated state machine. Each "machine" makes the same
changes to a data set so that each "machine" has the same "state."

The machine in our case is our `Addendum` object. The state changing
events will be the postbacks from Compassion which will arrive in the
same order for all participants.

The original participant will create the init with an initial data set.
There is a `bootstrap` event that we can respond to that will only be
invoked if the participant is the very first participant. The
bootstrapped participant can use this perform any initialization to the
data set. We don't need to respond to this event because out empty
database is simply an empty object used as a map.

Subsequent participants can receive a `join` event. We use the join
event in Addendum to fetch a copy of the database from one of the other
participants. It can't be just any copy. It can't be, say, the current
copy that any other participant has on hand. It must be the database
exactly as it was when we arrived, without any modifications made after
the participant arrive and not omitting any modifications made before it
arrived.

We use the `arrive` event to freeze a copy of the database when a new
participant arrives into a Cubbyhole. We will use that frozen copy of
the database specific to the promise (event id) of the participant's
arrival should the participant choose us when it asks for a copy of the
database when it joins.

This is a common operation, getting a backlog of work, or a snapshot of
work (hmm… maybe rename the postback from backlog to snapshot). An
Addendum participant makes its frozen copy of the database available
using the backlog postback. It provides a backlog end point available.
The joining participant calls this endpoint indirectly by calling
`./backlog` on the local Compassion API which will forward the request
to the leader. The Compassion API backlog endpoint can only be called
during the join postback.

Addendum responds to the acclimated event to clear out its frozen
snapshots. The acclimated event fires when an arriving participant has
completed its arrival. When the arrival for a participant is completed
we know that we do not have to continue to hold onto the frozen copy of
the database any longer so we free that memory.