Create a ping timeout helper.
When we do not receive a ping after a certain amount of time we are
supposed to terminate ourselves and not at all gracefully.

This helper class will check a timer and reset it occasionally.

Is is separate for purposes of testing and coverage. Also because when
we replay we're going to want to maybe turn off pinging. Otherwise, we
could just ping on occasion during replay to prevent the ping timeout.

Actually, I'm wondering about this as I create it. Why are we adding a
timer to the application when we're going to admonish the user to always
use the API to preserve a log for replay?

I suppose because network timeouts are part of the network layer. If
this was a persistent socket it would have a timeout built into it. We
would have to occasionally send a keep-alive message.