Implement departure by clearing cubbyholes.
On the off chance that we have a frozen copy of the database for a new
participant that drops out before they acclimate, we must delete that
frozen copy of the database. Cubbyhole will delete the entry if it
exists, do nothing if it does not.

Acclimate is an event that might never fire. It isn't a merely matter of
everyone receiving the arrival message. They arriving member must also
receive it. It is indicating that it has been initialized.

It does not then add a new message to the queue itself. It sets a flag
that propagates back to the leader eventually and then the leader sends
creates a new government that contains the acclimation record. The
propagation rids back on synchronize messages. We do this so that a
participant is not actively trying to postback to the leader essentially
duplicating all the logic that we've implemented at higher levels of
abstraction that handle submission and resubmission of messages.