Implement `add`.
In order to implement `add` I added a Turnstile to create a submission
queue for broadcast requests to the local API. The API is going to be
HTTP and the Reactor used to implement that is already going to queue
requests so the Turnstile for broadcasts should be as parallel as the
Turnstile for the HTTP requests.

Use Cliffhanger to wait on a reduced response before returning. Not
certain if this is precisely compatible with `etcd`. Does `etcd` return
once it knows the message is enqueued or once it knows the message has
reached all parties (i.e. it is consistent?) If return on enqueued we
need to respond from to the receive message. If return on consistent
then we leave it as is waiting on reduced message.