Track an index to use for `etcd`-like responses.
Update add to be able to return a response along the lines of what
`etcd` returns with some sort of indication of an index. Not sure what
that index is about. Assuming that the user is not supposed to make
assumptions about the value of the index beyond the fact that it
increases. We have that property and the fact that index values will be
in atomic log order. At some point let's look harder at the `etcd` index
and make sure our behavior covers its behavior.